# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Tapio Observability Platform API
  description: |
    Tapio provides enterprise-grade observability with semantic correlation.
    
    ## Features
    - Real-time event streaming with OTEL context
    - AI-powered semantic correlation
    - Multi-dimensional root cause analysis
    - Predictive analytics and business impact assessment
    
    ## Authentication
    API keys or JWT tokens in Authorization header
    
    ## Rate Limiting
    - 10,000 requests per minute per API key
    - 165,000 events per second for streaming
  version: 1.0.0
  contact:
    name: Tapio Support
    email: support@tapio.io
    url: https://tapio.io/support
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.tapio.io/api/v1
    description: Production server
  - url: https://staging-api.tapio.io/api/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /events/bulk:
    post:
      tags:
        - Events
      summary: Submit multiple events
      operationId: submitBulkEvents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/EventIngestRequest'
          application/x-ndjson:
            schema:
              type: string
              description: Newline-delimited JSON events
      responses:
        '202':
          description: Events accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkIngestResponse'
  /events/search:
    post:
      tags:
        - Events
      summary: Advanced event search
      operationId: searchEvents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventSearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventSearchResponse'
  /events/export:
    get:
      tags:
        - Events
      summary: Export events
      operationId: exportEvents
      parameters:
        - name: format
          in: query
          description: Export format
          schema:
            type: string
            enum: [json, csv, ndjson]
            default: json
        - name: start_time
          in: query
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Exported events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
            text/csv:
              schema:
                type: string
            application/x-ndjson:
              schema:
                type: string
...
  /correlations/patterns:
    post:
      tags:
        - Correlations
      summary: Discover patterns
      operationId: discoverPatterns
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatternDiscoveryRequest'
      responses:
        '200':
          description: Discovered patterns
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatternDiscoveryResponse'
  /correlations/impact:
    post:
      tags:
        - Correlations
      summary: Analyze event impact
      operationId: analyzeImpact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImpactAnalysisRequest'
      responses:
        '200':
          description: Impact analysis results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImpactAnalysisResponse'
...
  /analytics/summary:
    get:
      tags:
        - Analytics
      summary: Get analytics summary
      operationId: getAnalyticsSummary
      parameters:
        - name: start_time
          in: query
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsSummaryResponse'
  /analytics/trends:
    get:
      tags:
        - Analytics
      summary: Get trend analysis
      operationId: getTrendAnalysis
      parameters:
        - name: metric
          in: query
          schema:
            type: string
        - name: period
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Trend analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendAnalysisResponse'
...
components:
  schemas:
    EventIngestRequest:
      type: object
      required:
        - id
        - type
        - severity
        - timestamp
        - message
      properties:
        id:
          type: string
          example: evt_001
        type:
          type: string
          example: network
        severity:
          type: string
          enum: [debug, info, warning, error, critical]
          example: info
        timestamp:
          type: string
          format: date-time
        message:
          type: string
          example: Network connection established
        service:
          type: string
        component:
          type: string
        data:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
    EventIngestResponse:
      type: object
      properties:
        event_id:
          type: string
        status:
          type: string
          enum: [accepted, failed]
        error:
          type: string
        timestamp:
          type: string
          format: date-time
    BulkIngestResponse:
      type: object
      properties:
        total:
          type: integer
        success:
          type: integer
        failed:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/EventIngestResponse'
        timestamp:
          type: string
          format: date-time
    Event:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        severity:
          type: string
        timestamp:
          type: string
          format: date-time
        message:
          type: string
        context:
          type: object
          properties:
            service:
              type: string
            namespace:
              type: string
            trace_id:
              type: string
            span_id:
              type: string
    EventSearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
        time_range:
          $ref: '#/components/schemas/TimeRange'
        filters:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        limit:
          type: integer
        offset:
          type: integer
        sort_by:
          type: string
        sort_order:
          type: string
          enum: [asc, desc]
    EventSearchResponse:
      type: object
      properties:
        query:
          type: string
        total_hits:
          type: integer
          format: int64
        returned_hits:
          type: integer
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventSearchResult'
        facets:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/FacetValue'
        timestamp:
          type: string
          format: date-time
    EventSearchResult:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        severity:
          type: string
        timestamp:
          type: string
          format: date-time
        message:
          type: string
        score:
          type: number
        highlight:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    FacetValue:
      type: object
      properties:
        value:
          type: string
        count:
          type: integer
          format: int64
...
    PatternDiscoveryRequest:
      type: object
      properties:
        time_range:
          $ref: '#/components/schemas/TimeRange'
        min_confidence:
          type: number
        pattern_types:
          type: array
          items:
            type: string
        max_patterns:
          type: integer
    PatternDiscoveryResponse:
      type: object
      properties:
        time_range:
          $ref: '#/components/schemas/TimeRange'
        patterns:
          type: array
          items:
            $ref: '#/components/schemas/DiscoveredPattern'
        timestamp:
          type: string
          format: date-time
    DiscoveredPattern:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        confidence:
          type: number
        frequency:
          type: integer
        examples:
          type: array
          items:
            type: string
    ImpactAnalysisRequest:
      type: object
      required:
        - event_id
      properties:
        event_id:
          type: string
        include_services:
          type: boolean
        include_metrics:
          type: boolean
        time_horizon:
          type: string
    ImpactAnalysisResponse:
      type: object
      properties:
        event_id:
          type: string
        impact:
          $ref: '#/components/schemas/ImpactDetails'
        affected_services:
          type: array
          items:
            type: string
        affected_customers:
          type: integer
        estimated_duration:
          type: string
        recommendations:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
    ImpactDetails:
      type: object
      properties:
        business_impact:
          type: number
        customer_impact:
          type: number
        operational_impact:
          type: number
        financial_impact:
          type: number
...
    AnalyticsSummaryResponse:
      type: object
      properties:
        time_range:
          $ref: '#/components/schemas/TimeRange'
        event_statistics:
          $ref: '#/components/schemas/EventStats'
        correlation_statistics:
          $ref: '#/components/schemas/CorrelationStats'
        top_issues:
          type: array
          items:
            $ref: '#/components/schemas/TopIssue'
        timestamp:
          type: string
          format: date-time
    EventStats:
      type: object
      properties:
        total:
          type: integer
          format: int64
        by_type:
          type: object
          additionalProperties:
            type: integer
            format: int64
        by_severity:
          type: object
          additionalProperties:
            type: integer
            format: int64
        events_per_hour:
          type: array
          items:
            type: integer
            format: int64
    CorrelationStats:
      type: object
      properties:
        total:
          type: integer
          format: int64
        by_pattern:
          type: object
          additionalProperties:
            type: integer
            format: int64
        average_confidence:
          type: number
    TopIssue:
      type: object
      properties:
        description:
          type: string
        severity:
          type: string
        count:
          type: integer
        trend:
          type: string
    TrendAnalysisResponse:
      type: object
      properties:
        metric:
          type: string
        period:
          type: string
        trends:
          type: array
          items:
            $ref: '#/components/schemas/TrendData'
        prediction:
          $ref: '#/components/schemas/PredictionData'
        anomalies:
          type: array
          items:
            $ref: '#/components/schemas/AnomalyData'
        timestamp:
          type: string
          format: date-time
    TrendData:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        value:
          type: number
        trend:
          type: string
    PredictionData:
      type: object
      properties:
        next_value:
          type: number
        confidence:
          type: number
        trend:
          type: string
    AnomalyData:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        value:
          type: number
        description:
          type: string
        severity:
          type: string
...
    TimeRange:
      type: object
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT